workflow/requirements.txt
Django==2.2.3
django-pagedown==1.0.6
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
